untrack a couple of nested files
These are in js/ and ruby/, the contents of which are supposed to be
ignored.  They must have gotten committed before the ignore, so we
untrack them here.

The $64 question:  Does this commit remove them from an otherwise
up-to-date repo, or are they protected by a previous commit that ignores
them?